memory: store item instead of serialized data
Previously stored the serialized data so it needed to be deserialized
to check if it has expired. ie. GC needs to deserialze the whole corpus.

This change updates the code to store the *onecache.Item. The ExpiresAt
can then be used directly to check expires.

During Set calls the data is copied before save and during Get calls
the data is copied before return. This allows the caller to modify their
data without issue, as they were allowed to before.